+++ NEWS.md
@@ -1,4 +1,4 @@
-# lavaan.printer 0.0.0.9003
+# lavaan.printer 0.0.0.9004
 
 - Add `est_funs_args`, `header_funs_args`,
   and `footer_funs_args`. (0.0.0.9001)
@@ -7,6 +7,10 @@
 
 - Updated the documentation and examples.
   (0.0.0.9003)
+
+- Add support for line wrapping when
+  a section is a character vector,
+  usually a list of notes. (0.0.0.9004)
 
 # lavaan.printer 0.0.0.9000